+++ _Typings/gameApi/types/collider.d.ts
@@ -23,8 +23,8 @@ declare module "game:type" {
 
     type CollisionEvent = {
         eventType: EventType
-        itemA: Item | Player
-        itemB: Item | Player
+        itemA: Item
+        itemB: Item
         impulse: float
         contactPoint: Float3
         normal: Float3